+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator-list/indicator-list.adapter.ts
