+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator-list/mapper/indicator-list.mapper.ts
