+++ api/src/numerical-guidance/numerical-guidance.module.ts
@@ -6,12 +6,12 @@ import { FluctuatingIndicatorKrxAdapter } from './infrastructure/adapter/krx/flu
 import { CqrsModule } from '@nestjs/cqrs';
 import { HttpModule } from '@nestjs/axios';
 import { GetIndicatorListQueryHandler } from './application/query/get-indicator-list/get-indicator-list.query.handler';
-import { IndicatorListAdapter } from './infrastructure/adapter/indicator-list/indicator-list.adapter';
+import { IndicatorListAdapter } from './infrastructure/adapter/persistence/indicator-list/indicator-list.adapter';
 import { GetFluctuatingIndicatorWithoutCacheQueryHandler } from './application/query/get-fluctuatingIndicator-without-cache/get-fluctuatingIndicator-without-cache.query.handler';
 import { CreateIndicatorBoardMetadataCommandHandler } from './application/command/create-indicator-board-metadata/create-indicator-board-metadata.command.handler';
-import { IndicatorBoardMetadataPersistentAdapter } from './infrastructure/adapter/persistence/indicator-board-metadata.persistent.adapter';
+import { IndicatorBoardMetadataPersistentAdapter } from './infrastructure/adapter/persistence/indicator-board-metadata/indicator-board-metadata.persistent.adapter';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { IndicatorBoardMetadataEntity } from './infrastructure/adapter/persistence/entity/indicator-board-metadata.entity';
+import { IndicatorBoardMetadataEntity } from './infrastructure/adapter/persistence/indicator-board-metadata/entity/indicator-board-metadata.entity';
 import { AuthService } from '../auth/auth.service';
 import { MemberEntity } from '../auth/member.entity';
 import { GetIndicatorBoardMetaDataQueryHandler } from './application/query/get-indicator-board-metadata/get-indicator-board-metadata.query.handler';
